Add assemblyresolver to console project too + other MR feedback
