+++ Application/Riskeer/src/Application.Riskeer.Migration.Console/app.config
@@ -22,9 +22,4 @@ Stichting Deltares and remain full property of Stichting Deltares at all times.
 All rights reserved.
 -->
 <configuration>
-    <runtime>
-        <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-            <probing privatePath="Application;Application\Built-in;Application\Built-in\Managed;Application\Built-in\Managed\Core;Application\Built-in\Managed\Deltares;Application\Built-in\Managed\Deltares\AssemblyTool-23.1.1;Application\Built-in\Managed\Deltares\D-Stability-20.2.1;Application\Built-in\Managed\Deltares\Macrostability-22.1.1;Application\Built-in\Managed\Deltares\Piping-16.2.1.4574;Application\Built-in\Managed\Third-party;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\de;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\es;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\fr;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\hu;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\it;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\nl;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\pt-BR;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\ro;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\ru;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\sv;Application\Built-in\Managed\Third-party\AvalonDock-2.0.2000\zh-Hans;Application\Built-in\Managed\Third-party\BruTile-3.1.3;Application\Built-in\Managed\Third-party\BruTile.Desktop-3.1.3;Application\Built-in\Managed\Third-party\ControlzEx-4.4.0;Application\Built-in\Managed\Third-party\DotSpatial.Controls-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Controls-2.0.0-rc1\de-DE;Application\Built-in\Managed\Third-party\DotSpatial.Controls-2.0.0-rc1\it-IT;Application\Built-in\Managed\Third-party\DotSpatial.Controls-2.0.0-rc1\pt-BR;Application\Built-in\Managed\Third-party\DotSpatial.Controls-2.0.0-rc1\zh-CN;Application\Built-in\Managed\Third-party\DotSpatial.Data-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Data.Forms-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Data.Forms-2.0.0-rc1\zh-CN;Application\Built-in\Managed\Third-party\DotSpatial.Extensions-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.GeoAPI-1.7.4.3;Application\Built-in\Managed\Third-party\DotSpatial.Modeling.Forms-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Modeling.Forms-2.0.0-rc1\zh-CN;Application\Built-in\Managed\Third-party\DotSpatial.Modeling.Forms-2.0.0-rc1\zh-MO;Application\Built-in\Managed\Third-party\DotSpatial.Mono-1.9.0;Application\Built-in\Managed\Third-party\DotSpatial.NetTopologySuite-1.14.4;Application\Built-in\Managed\Third-party\DotSpatial.NTSExtension-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Projections-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\cs;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\de;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\de-DE;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\el;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\it-IT;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\ja-JP;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\lt;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\nl;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\pt;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\pt-BR;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\ro;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\th;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\vi-VN;Application\Built-in\Managed\Third-party\DotSpatial.Projections.Forms-2.0.0-rc1\zh-CN;Application\Built-in\Managed\Third-party\DotSpatial.Serialization-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Symbology-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Symbology-2.0.0-rc1\de-DE;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\cs;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\de-DE;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\en-US;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\it;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\it-IT;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\pt-BR;Application\Built-in\Managed\Third-party\DotSpatial.Symbology.Forms-2.0.0-rc1\zh-CN;Application\Built-in\Managed\Third-party\DotSpatial.Topology-1.9.0;Application\Built-in\Managed\Third-party\EntityFramework-6.4.4;Application\Built-in\Managed\Third-party\FontAwesome.Sharp-5.15.3;Application\Built-in\Managed\Third-party\GeoAPI-1.7.4;Application\Built-in\Managed\Third-party\GraphSharp-1.1.0;Application\Built-in\Managed\Third-party\Log4net-3.3.1;Application\Built-in\Managed\Third-party\MahApps.Metro-2.4.4;Application\Built-in\Managed\Third-party\MahApps.Metro-2.4.4\de;Application\Built-in\Managed\Third-party\MathNet.Numerics-3.19.0;Application\Built-in\Managed\Third-party\MathNet.Spatial-0.3.0;Application\Built-in\Managed\Third-party\Microsoft.Bcl.HashCode-6.0.0;Application\Built-in\Managed\Third-party\Microsoft.Xaml.Behaviors.Wpf-1.1.31;Application\Built-in\Managed\Third-party\Microsoft.Xaml.Behaviors.Wpf-1.1.31\Design;Application\Built-in\Managed\Third-party\NetTopologySuite-1.14.0;Application\Built-in\Managed\Third-party\OxyPlot.Core-1.0.0;Application\Built-in\Managed\Third-party\OxyPlot.WindowsForms-1.0.0;Application\Built-in\Managed\Third-party\QuickGraph-3.6.61119.7;Application\Built-in\Managed\Third-party\QuickGraph-3.6.61119.7\CodeContracts;Application\Built-in\Managed\Third-party\SmartThreadPool.dll-2.2.4;Application\Built-in\Managed\Third-party\System.Buffers-4.6.1;Application\Built-in\Managed\Third-party\System.Collections.Immutable-10.0.5;Application\Built-in\Managed\Third-party\System.Data.SQLite.Core-1.0.117;Application\Built-in\Managed\Third-party\System.Data.SQLite.Core-1.0.117\x64;Application\Built-in\Managed\Third-party\System.Data.SQLite.Core-1.0.117\x86;Application\Built-in\Managed\Third-party\System.Data.SQLite.EF6-1.0.117;Application\Built-in\Managed\Third-party\System.Formats.Nrbf-10.0.5;Application\Built-in\Managed\Third-party\System.Memory-4.6.3;Application\Built-in\Managed\Third-party\System.Numerics.Vectors-4.6.1;Application\Built-in\Managed\Third-party\System.Reflection.Metadata-10.0.5;Application\Built-in\Managed\Third-party\System.Resources.Extensions-10.0.5;Application\Built-in\Managed\Third-party\System.Runtime.CompilerServices.Unsafe-6.1.2;Application\Built-in\Managed\Third-party\System.ValueTuple-4.6.1;Application\Built-in\Managed\Third-party\WPFExtensions-1.0.0;" />
-        </assemblyBinding>
-    </runtime>
 </configuration>
