+++ Application/Riskeer/src/Application.Riskeer/App.xaml.cs
@@ -29,7 +29,6 @@ using System.IO;
 using System.IO.Packaging;
 using System.Linq;
 using System.Reflection;
-using System.Runtime.Loader;
 using System.Threading;
 using System.Windows;
 using System.Windows.Controls.Primitives;
@@ -75,16 +74,22 @@ namespace Application.Riskeer
 
         private static string fileToOpen = string.Empty;
 
-        private readonly ILog log;
-        private GuiCore gui;
-
-        private delegate void ExceptionDelegate(Exception exception);
-        
-        // Create a static instance of the App class to ensure that the AssemblyResolve event is subscribed to before any assemblies are loaded.
+        /// <summary>
+        /// Creates a static instance of <see cref="App"/>.
+        /// </summary>
+        /// <remarks>
+        /// This constructor is used to subscribe <see cref="AppDomain.AssemblyResolve"/> before any assemblies are loaded.
+        /// </remarks>
         static App()
         {
             AppDomain.CurrentDomain.AssemblyResolve += AssemblyResolver.ResolveAssembly;
         }
+
+        private readonly ILog log;
+        private GuiCore gui;
+
+        private delegate void ExceptionDelegate(Exception exception);
+
         /// <summary>
         /// Creates a new instance of <see cref="App"/>.
         /// </summary>
