+++ Application/Riskeer/src/Application.Riskeer/Application.Riskeer.csproj
@@ -72,6 +72,11 @@ All rights reserved.
     <ProjectReference Include="..\..\..\..\Riskeer\StabilityStoneCover\src\Riskeer.StabilityStoneCover.Plugin\Riskeer.StabilityStoneCover.Plugin.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\..\build\AssemblyResolver.cs">
+        <Link>AssemblyResolver.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
     <None Include="..\..\..\..\build\Copying.licenseheader">
       <Link>Copying.licenseheader</Link>
     </None>
@@ -111,9 +116,6 @@ All rights reserved.
       <Version>1.0.117</Version>
       <GeneratePathProperty>true</GeneratePathProperty>
       <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-    <PackageReference Include="System.Runtime.Loader">
-      <Version>4.3.0</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
