+++ Application/Riskeer/src/Application.Riskeer/app.config
@@ -56,9 +56,4 @@ All rights reserved.
       <add name="SQLite Data Provider (Entity Framework 6)" invariant="System.Data.SQLite.EF6" description=".NET Framework Data Provider for SQLite (Entity Framework 6)" type="System.Data.SQLite.EF6.SQLiteProviderFactory, System.Data.SQLite.EF6" />
     </DbProviderFactories>
   </system.data>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <probing privatePath="Application;Application\Built-in;" />
-    </assemblyBinding>
-  </runtime>
 </configuration>
