+++ Application/Riskeer/test/Application.Riskeer.Integration.Test/Application.Riskeer.Integration.Test.csproj
@@ -56,6 +56,11 @@ All rights reserved.
     <ProjectReference Include="..\..\src\Application.Riskeer\Application.Riskeer.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\..\build\AssemblyResolver.cs">
+      <Link>AssemblyResolverTest/AssemblyResolver.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
     <PackageReference Include="MahApps.Metro">
       <Version>2.4.4</Version>
     </PackageReference>
@@ -75,19 +80,6 @@ All rights reserved.
       <Version>1.0.117</Version>
     </PackageReference>
   </ItemGroup>
-  <!-- Copy the managed assemblies to the test output directory for AssemblyResolverTest -->
-  <ItemGroup>
-    <None Include="$(SolutionDir)bin\Debug\Application\Built-in\Managed\**\*">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <Link>Application\Built-in\Managed\%(RecursiveDir)%(Filename)%(Extension)</Link>
-    </None>
-  </ItemGroup>
-  <Target Name="CopyManagedAssemblies" AfterTargets="Build">
-    <ItemGroup>
-      <ManagedAssemblies Include="$(SolutionDir)bin\Debug\Application\Built-in\Managed\**\*" />
-    </ItemGroup>
-    <Copy SourceFiles="@(ManagedAssemblies)" DestinationFiles="@(ManagedAssemblies->'$(OutDir)Application\Built-in\Managed\%(RecursiveDir)%(Filename)%(Extension)')" SkipUnchangedFiles="true" />
-  </Target>
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
     <Exec Command="xcopy &quot;$(ProjectDir)UnusedResourceSearcher&quot; &quot;$(OutDir)UnusedResourceSearcher\*&quot; /D /S /Y" />
   </Target>
