+++ Migration/Console/src/Migration.Console/Migration.Console.csproj
@@ -27,6 +27,11 @@ All rights reserved.
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\..\build\AssemblyResolver.cs">
+      <Link>AssemblyResolver.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
     <None Include="..\..\..\..\build\Copying.Lesser.licenseheader">
       <Link>Copying.Lesser.licenseheader</Link>
     </None>
